Fixed problem with python 3 builds to test for a different initial Faker.name() value.  Apparently the random number generator for python 2 and python 3 create different sequences with the same seed.
This is new to me and I had no idea that the project would be
building on different versions of python nor that python 3 and
python 2 would have different seeding sequences inside the random
number generator.  I was able to set up virtual environments for
all of the different python versions and test the code in all of
them so I won't break the builds any more if I check in something
else.